requirements.txt
Flask
Flask-SocketIO
Flask-Cors
PyYAML
opencv-python-headless
openpyxl
eventlet
